Added remote folder option (NFS performance optimization)
Runing statparse on an NFS-share was very slow. The reason is was that
starting the python interpreter in a mounted directory takes upwards of
3-5 seconds. However, loading a small experiments takes around 0.3s
which is about 0.05s slower than running the script locally.

The fix is to start the python interpreter in a local directory and then
switch to the remote directory when needed. The same optimization can
likely be applied to other scripts as well.