Regression tests and bugfix
Implemented three regression tests to ensure that the model works as
intended. In this process, I found a bug in the static energy model for
the parallel case where I erroneously multiplied by the number of cores.
This is incorrect because it fails to take into account that although we
do power N times as many cores, the runtime is decreased with a factor
of N. Thus, the static energy is not affected by N.