New and validated power model
Added the new and validated power model from Frode and updated the test
cases. Changed the way to specify the work per period by specifying it
as the fraction of maximum performance of a single core. Also removed
the power model from the scripts since it is much simpler to compute the
static and dynamic power consumption per core for each operating point
in Excel.